Move util into its own package for better reusability
