+++ controllers/auth.go
@@ -8,6 +8,7 @@ import (
 	"github.com/steffen25/golang.zone/app"
 	"github.com/steffen25/golang.zone/repositories"
 	"github.com/steffen25/golang.zone/services"
+	"github.com/steffen25/golang.zone/util"
 )
 
 type AuthController struct {
@@ -31,7 +32,7 @@ func (ac *AuthController) Authenticate(w http.ResponseWriter, r *http.Request) {
 		NewAPIError(&APIError{false, "Email is required", http.StatusBadRequest}, w)
 		return
 	}
-	if ok := IsEmail(email); !ok {
+	if ok := util.IsEmail(email); !ok {
 		NewAPIError(&APIError{false, "You must provide a valid email address", http.StatusBadRequest}, w)
 		return
 	}
