+++ util/util.go
@@ -1,4 +1,4 @@
-package controllers
+package util
 
 import (
 	"regexp"
@@ -14,7 +14,7 @@ func IsEmail(email string) bool {
 	return true
 }
 
-func generateSlug(title string) string {
+func GenerateSlug(title string) string {
 	re := regexp.MustCompile("[^a-z0-9]+")
 	return strings.Trim(re.ReplaceAllString(strings.ToLower(title), "-"), "-")
 }
